+++ src/main/java/com/web/cntt/model/GiangVien.java
@@ -26,7 +26,7 @@ public class GiangVien extends BaseEntity {
     private String ten;
     private Date ngayVL;
     private String chucVu;
-    private int heSoLuong;
+    private double heSoLuong;
     private String hocVi;
     private String gioiTinh;
     private String sdt;
